refactor: app.dart, main.dart 합치기 DRONI-103
기존의 분리된 코드를 main.dart에 합쳤습니다.
