+++ lib/main.dart
@@ -1,8 +1,12 @@
+import 'package:droni/common/dart/extension/context_extension.dart';
+import 'package:droni/common/theme/custom_theme.dart';
+import 'package:droni/common/theme/custom_theme_app.dart';
+import 'package:droni/screen/login/s_login.dart';
 import 'package:flutter/material.dart';
 
 import 'package:easy_localization/easy_localization.dart';
+import 'package:nav/nav.dart';
 
-import 'app.dart';
 import 'common/data/preference/app_preferences.dart';
 
 void main() async {
@@ -19,4 +23,71 @@ void main() async {
       child: const App(),
     ),
   );
+}
+
+class App extends StatefulWidget {
+  static final GlobalKey<NavigatorState> navigatorKey = GlobalKey();
+
+  ///light, dark 테마가 준비되었고, 시스템 테마를 따라가게 하려면 해당 필드를 제거 하시면 됩니다.
+  static const defaultTheme = CustomTheme.light;
+  static bool isForeground = true;
+
+  const App({super.key});
+
+  @override
+  State<App> createState() => AppState();
+}
+
+class AppState extends State<App> with Nav, WidgetsBindingObserver {
+  @override
+  GlobalKey<NavigatorState> get navigatorKey => App.navigatorKey;
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addObserver(this);
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomThemeApp(
+      child: Builder(builder: (context) {
+        return MaterialApp(
+          navigatorKey: App.navigatorKey,
+          localizationsDelegates: context.localizationDelegates,
+          supportedLocales: context.supportedLocales,
+          debugShowCheckedModeBanner: false,
+          locale: context.locale,
+          title: 'Image Finder',
+          theme: context.themeType.themeData,
+          home: const LoginScreen(),
+        );
+      }),
+    );
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    switch (state) {
+      case AppLifecycleState.resumed:
+        App.isForeground = true;
+        break;
+      case AppLifecycleState.inactive:
+        break;
+      case AppLifecycleState.paused:
+        App.isForeground = false;
+        break;
+      case AppLifecycleState.detached:
+        break;
+      case AppLifecycleState.hidden: //Flutter 3.13 이하 버전을 쓰신다면 해당 라인을 삭제해주세요.
+        break;
+    }
+    super.didChangeAppLifecycleState(state);
+  }
 }
